+++ CHANGELOG.md
@@ -10,6 +10,8 @@ Note: In this file, do not use the hard wrap in the middle of a sentence for com
 
 ## [Unreleased]
 
+## [0.1.16] - 2024-05-24
+
 - Add `TargetSpec::link_self_contained`.
 
 ## [0.1.15] - 2024-04-07
@@ -82,7 +84,8 @@ Note: In this file, do not use the hard wrap in the middle of a sentence for com
 
 Initial release
 
-[Unreleased]: https://github.com/taiki-e/target-spec-json/compare/v0.1.15...HEAD
+[Unreleased]: https://github.com/taiki-e/target-spec-json/compare/v0.1.16...HEAD
+[0.1.16]: https://github.com/taiki-e/target-spec-json/compare/v0.1.15...v0.1.16
 [0.1.15]: https://github.com/taiki-e/target-spec-json/compare/v0.1.14...v0.1.15
 [0.1.14]: https://github.com/taiki-e/target-spec-json/compare/v0.1.13...v0.1.14
 [0.1.13]: https://github.com/taiki-e/target-spec-json/compare/v0.1.12...v0.1.13